Load an atomic vector

Load a simple vector consisting of atomic elements from file.

loadAtomicVector(info, project, ...)

Arguments

info

Named list containing the metadata for this object.

project

Any argument accepted by the acquisition functions, see ?acquireFile. By default, this should be a string containing the path to a staging directory.

...

Further arguments, ignored.

Value

The vector described by info, possibly with names.

See also

Author

Aaron Lun

Examples

tmp <- tempfile()
dir.create(tmp)
meta <- stageObject(setNames(runif(26), letters), tmp, path="bar")
loadAtomicVector(meta, tmp)
#>          a          b          c          d          e          f          g 
#> 0.30023820 0.32305066 0.02231284 0.85630217 0.02890491 0.09624149 0.64310236 
#>          h          i          j          k          l          m          n 
#> 0.24286169 0.20905423 0.77998908 0.51360842 0.23577732 0.13864953 0.76635888 
#>          o          p          q          r          s          t          u 
#> 0.83408448 0.45810244 0.46825863 0.28132679 0.88083471 0.17438886 0.95124070 
#>          v          w          x          y          z 
#> 0.39964047 0.02164273 0.96770380 0.80353722 0.12873304