Load a base list

Load a list from file, possibly containing complex entries.

loadBaseList(info, project, parallel = TRUE)

Arguments

info

Named list containing the metadata for this object.

project

Any argument accepted by the acquisition functions, see ?acquireFile. By default, this should be a string containing the path to a staging directory.

parallel

Whether to perform reading and parsing in parallel for greater speed. Only relevant for lists stored in the JSON format.

Value

The list described by info.

Details

This function effectively reverses the behavior of "stageObject,list-method", loading the list back into memory from the JSON file. Atomic vectors, arrays and data frames are loaded directly while complex values are loaded by calling the appropriate loading function.

See also

Author

Aaron Lun

Examples

library(S4Vectors)
ll <- list(A=1, B=LETTERS, C=DataFrame(X=letters))

# First staging it:
tmp <- tempfile()
dir.create(tmp)
info <- stageObject(ll, tmp, path="stuff")

# And now loading it:
loadBaseList(info, tmp)
#> $A
#> [1] 1
#> 
#> $B
#>  [1] "A" "B" "C" "D" "E" "F" "G" "H" "I" "J" "K" "L" "M" "N" "O" "P" "Q" "R" "S"
#> [20] "T" "U" "V" "W" "X" "Y" "Z"
#> 
#> $C
#> DataFrame with 26 rows and 1 column
#>               X
#>     <character>
#> 1             a
#> 2             b
#> 3             c
#> 4             d
#> 5             e
#> ...         ...
#> 22            v
#> 23            w
#> 24            x
#> 25            y
#> 26            z
#>