Load an object from its metadata

Load an object from its metadata, based on the loading function specified in its schema.

loadObject(info, project, ...)

Arguments

info

Named list containing the metadata for this object.

project

Any argument accepted by the acquisition functions, see ?acquireFile. By default, this should be a string containing the path to a staging directory.

...

Further arguments to pass to the specific loading function listed in the schema.

Value

An object corresponding to info, as defined by the loading function.

Details

The loadObject function is intended to load objects where the class is not known in advance. Most typically, this is indirectly called inside other loading functions to restore child objects of arbitrary type. Once in memory, the child objects can then be assembled into more complex objects by the caller. (It would be unwise to use loadObject to try restore a non-child object as this would result in infinite recursion.)

This function will look through the schemas in alabaster.schemas to find the schema specified in info$`$schema`. Upon discovery, loadObject will extract the loading function from the _attributes.restore.R property of the schema; this should be a string that contains a namespaced function, which can be parsed and evaluated to obtain said function. loadObject will then call the loading function with the supplied arguments. Developers can temporarily add extra packages to the schema search path by supplying package names in the alabaster.schema.locations option; schema files are expected to be stored in the schemas subdirectory of each package's installation directory.

Developers of Artificer extensions should use .loadObject rather than calling loadObject directly. This ensures that any application-level overrides of the loading functions are respected. Developers of Artificer applications should also read the commentary in ?".altLoadObject".

Author

Aaron Lun

Examples

# Same example as stageObject, but reversed.
library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])

# First staging it:
tmp <- tempfile()
dir.create(tmp)
out <- stageObject(df, tmp, path="coldata")

# Loading it:
loadObject(out, tmp)
#> DataFrame with 10 rows and 2 columns
#>            A           B
#>    <integer> <character>
#> 1          1           A
#> 2          2           B
#> 3          3           C
#> 4          4           D
#> 5          5           E
#> 6          6           F
#> 7          7           G
#> 8          8           H
#> 9          9           I
#> 10        10           J