HDF5 storage type optimization

Optimize HDF5 storage to use the smallest possible storage type. Intended for alabaster.* developers only.

collect_integer_attributes(x)

optimize_integer_storage(x)

collect_number_attributes(x)

optimize_number_storage(x, fallback = chooseMissingPlaceholderForHdf5)

collect_string_attributes(x)

optimize_string_storage(x, fallback = NULL)

collect_boolean_attributes(x)

optimize_boolean_storage(x)

Arguments

x

An atomic vector or array of the specified type, e.g., integer vector/array for collect_integer_attributes and optimize_integer_storage. Developers can also extend this to abstract objects containing values of the same type, e.g., matrix-like S4 classes.

fallback

Function that accepts x and returns a suitable placeholder in the presence of NAs. If NULL, this defaults to chooseMissingPlaceholderForHdf5.

Value

For collect_integer_attributes, a named list containing:

min, the smallest non-NA value in x. This is set to Inf if all values are NA.

max, the largest non-NA value in x. This is set to -Inf if all values are NA.

missing, logical scalar indicating whether any values in x are NA.

For collect_number_attributes, a named list containing:

missing, logical scalar indicating whether any values in x are NA.

non_integer, logical scalar indicating whether any values in x are non-integer.

min, the smallest non-NA integer value in x. This is set to Inf if there are any non-integer or NA values.

max, the largest non-NA integer value in x. This is set to -Inf if there are any non-integer or NA values.

has_NaN, logical scalar indicating whether NaN is present in x.

has_Inf, logical scalar indicating whether positive infinity is present in x.

has_NegInf, logical scalar indicating whether negative infinity is present in x.

has_lowest, logical scalar indicating whether the smallest double-precision value is present in x.

has_highest, logical scalar indicating whether the highest double-precision value is present in x.

For collect_string_attributes, a named list containing:

missing, logical scalar indicating whether any values in x are NA.

has_NA, logical scalar indicating whether the "NA" string is present in x.

has__NA, logical scalar indicating whether the "_NA" string is present in x.

max_len, integer scalar specifying the maximum length of the strings in x.

For collect_boolean_attributes, a named list containing:

missing, logical scalar indicating whether any values in x are NA.

For the optimize_*_storage functions, a named list containing:

type, string containing the HDF5 datatype for storing x.

placeholder, value of the placeholder for NA values.

other, other attributes of x (e.g., number of non-zero elements for sparse vectors). These should be stored in an other field in the named list returned by collect_*_ attributes.

Author

Aaron Lun