Read an atomic vector from disk

Read a vector consisting of atomic elements from its on-disk representation.

readAtomicVector(path, ...)

Arguments

path

Path to a directory created with any of the vector methods for saveObject.

...

Further arguments, ignored.

Value

The vector described by info.

See also

Author

Aaron Lun

Examples

tmp <- tempfile()
saveObject(setNames(runif(26), letters), tmp)
readAtomicVector(tmp)
#>          a          b          c          d          e          f          g 
#> 0.70087920 0.46223692 0.70362507 0.41211430 0.70388587 0.51252195 0.58876318 
#>          h          i          j          k          l          m          n 
#> 0.65782672 0.29994023 0.45700319 0.92113038 0.98121844 0.92543565 0.38074062 
#>          o          p          q          r          s          t          u 
#> 0.52569097 0.47843552 0.59498188 0.09197069 0.44459837 0.17565265 0.12958471 
#>          v          w          x          y          z 
#> 0.97556154 0.60586312 0.59163752 0.98977043 0.01100330