Stage a DataFrame

Stage a DataFrame by saving it to a CSV or HDF5 file. CSV files follow the comservatory specification, while the expected layout of a HDF5 file is described in the hdf5_data_frame schema in alabaster.schemas.

# S4 method for DataFrame
stageObject(
  x,
  dir,
  path,
  child = FALSE,
  df.name = "simple",
  mcols.name = "mcols",
  meta.name = "other"
)

Arguments

x

A DataFrame.

dir

String containing the path to the staging directory.

path

String containing a prefix of the relative path inside dir where x is to be saved. The actual path used to save x may include additional components, see Details.

child

Logical scalar indicating whether x is a child of a larger object.

df.name

String containing the relative path inside dir to save the CSV/HDF5 file.

mcols.name

String specifying the name of the directory inside path to save mcols(x). If NULL, per-element metadata is not saved.

meta.name

String specifying the name of the directory inside path to save metadata(x). If NULL, object metadata is not saved.

Value

A named list containing the metadata for x. x itself is written to a CSV or HDF5 file inside path. Additional files may also be created inside path and referenced from the metadata.

Details

All atomic vector types are supported in the columns along with dates and (ordered) factors. Dates and factors are converted to character vectors and saved as such inside the file. Factor levels are saved in a separate data frame, which is referenced in the columns field of the returned metadata.

Any non-atomic columns are saved to a separate file inside path via stageObject, and referenced from the corresponding columns entry. For consistency, they will be replaced in the main file by a placeholder all-zero column.

As a DataFrame is a Vector subclass, its R-level metadata can be staged by .processMetadata.

File formats

If .saveDataFrameFormat() == "csv", the contents of x are saved to a uncompressed CSV file. If x has non-NULL row names, the first saved column in the CSV is named row_names and will contain the row names. This should be ignored when indexing columns and comparing them to the corresponding entry of columns in the file's JSON metadata document.

If .saveDataFrameFormat() == "csv.gz", the CSV file is compressed (the default). This reduces space and bandwidth requirements at the cost of the (de)compression overhead. It also makes it more difficult to do queries inside the file without decompression of the entire file.

If .saveDataFrameFormat() == "hdf5", x is saved into a HDF5 file instead of a CSV. Columns are saved into a data group where each column is a dataset named after its positional index. The names of the columns are saved into the column_names dataset. If row names are present, a separate row_names dataset containing the row names will be generated. This format is most useful for random access and for preserving the precision of numerical data.

See also

Author

Aaron Lun

Examples

library(S4Vectors)
df <- DataFrame(A=1:10, B=LETTERS[1:10])

tmp <- tempfile()
dir.create(tmp)
stageObject(df, tmp, path="coldata")
#> $`$schema`
#> [1] "csv_data_frame/v1.json"
#> 
#> $path
#> [1] "coldata/simple.csv.gz"
#> 
#> $is_child
#> [1] FALSE
#> 
#> $data_frame
#> $data_frame$columns
#> $data_frame$columns[[1]]
#> $data_frame$columns[[1]]$name
#> [1] "A"
#> 
#> $data_frame$columns[[1]]$type
#> [1] "integer"
#> 
#> 
#> $data_frame$columns[[2]]
#> $data_frame$columns[[2]]$name
#> [1] "B"
#> 
#> $data_frame$columns[[2]]$type
#> [1] "string"
#> 
#> 
#> 
#> $data_frame$row_names
#> [1] FALSE
#> 
#> $data_frame$column_data
#> NULL
#> 
#> $data_frame$other_data
#> NULL
#> 
#> $data_frame$dimensions
#> [1] 10  2
#> 
#> 
#> $csv_data_frame
#> $csv_data_frame$compression
#> [1] "gzip"
#> 
#> 

list.files(tmp, recursive=TRUE)
#> [1] "coldata/simple.csv.gz"