Stage a DataFrameFactor object

Stage a DataFrameFactor object, a generalization of the base factor for DataFrame levels.

# S4 method for DataFrameFactor
stageObject(
  x,
  dir,
  path,
  child = FALSE,
  index.name = "index",
  level.name = "levels",
  mcols.name = "mcols"
)

Arguments

x

A DataFrameFactor object.

dir

String containing the path to the staging directory.

path

String containing a prefix of the relative path inside dir where x is to be saved. The actual path used to save x may include additional components, see Details.

child

Logical scalar indicating whether x is a child of a larger object.

index.name

String containing the name of the file to save the factor indices.

level.name

String containing the name of the subdirectory to save the factor levels.

mcols.name

String specifying the name of the directory inside path to save the mcols. If NULL, the metadata columns are not saved.

Value

A named list containing the metadata for x. x itself is written to a file inside path. Additional files may also be created inside path and referenced from the metadata.

Details

We create one file in path for the levels and another file for the factor indices. The levels file contains a DataFrame as staged by stageObject,DataFrame-method. Indices are 1-based and reference one record of the levels file. As the DataFrameFactor is a Vector subclass, its R-level metadata can be staged by .processMetadata.

See also

Author

Aaron Lun

Examples

library(S4Vectors)
df <- DataFrame(X=LETTERS[1:5], Y=1:5)
out <- DataFrameFactor(df[sample(5, 100, replace=TRUE),,drop=FALSE])

tmp <- tempfile()
dir.create(tmp)
stageObject(out, tmp, path="test")
#> $`$schema`
#> [1] "data_frame_factor/v1.json"
#> 
#> $path
#> [1] "test/index"
#> 
#> $is_child
#> [1] FALSE
#> 
#> $factor
#> $factor$length
#> [1] 100
#> 
#> $factor$names
#> [1] FALSE
#> 
#> $factor$element_data
#> NULL
#> 
#> $factor$compression
#> [1] "gzip"
#> 
#> 
#> $data_frame_factor
#> $data_frame_factor$levels
#> $data_frame_factor$levels$resource
#> $data_frame_factor$levels$resource$type
#> [1] "local"
#> 
#> $data_frame_factor$levels$resource$path
#> [1] "test/levels/simple.csv.gz"
#> 
#> 
#> 
#> 

list.files(file.path(tmp, "test"), recursive=TRUE)
#> [1] "index"                     "levels/simple.csv.gz"     
#> [3] "levels/simple.csv.gz.json"