Read an atomic vector from disk

Read a vector consisting of atomic elements from its on-disk representation.

readAtomicVector(path, metadata, ...)

Arguments

path

Path to a directory created with any of the vector methods for saveObject.

metadata

Named list containing metadata for the object, see readObjectFile for details.

...

Further arguments, ignored.

Value

The vector described by info.

See also

Author

Aaron Lun

Examples

tmp <- tempfile()
saveObject(setNames(runif(26), letters), tmp)
readAtomicVector(tmp)
#>          a          b          c          d          e          f          g 
#> 0.10505917 0.26759438 0.00649745 0.29395229 0.61729847 0.88277415 0.96144142 
#>          h          i          j          k          l          m          n 
#> 0.78493664 0.61559544 0.97323696 0.27171829 0.95672349 0.25807627 0.66541884 
#>          o          p          q          r          s          t          u 
#> 0.50051972 0.94800500 0.92137500 0.91267127 0.70269400 0.02175440 0.98019865 
#>          v          w          x          y          z 
#> 0.46734377 0.98043363 0.27174433 0.03918217 0.94081707