Stage assorted objects

Generic to stage assorted R objects. More methods may be defined by other packages to extend the alabaster.base framework to new classes.

stageObject(x, dir, path, child = FALSE, ...)

Arguments

x

A Bioconductor object of the specified class.

dir

String containing the path to the staging directory.

path

String containing a prefix of the relative path inside dir where x is to be saved. The actual path used to save x may include additional components, see Details.

child

Logical scalar indicating whether x is a child of a larger object.

...

Further arguments to pass to specific methods.

Value

dir is populated with files containing the contents of x. All created files should be prefixed by path, either in the file name or as part of a subdirectory path.

A named list containing the metadata for x is returned, containing at least:

$schema, a string specifying the schema to use to validate the metadata. This may be decorated with the package attribute to help .writeMetadata find the package containing the schema.

path, a string containing the path to some file prefixed by the input path.

is_child, a logical scalar equal to the input child.

All files created by a stageObject method should be referenced from the metadata list, directly or otherwise (e.g., via child resources).

Details

Methods for the stageObject generic are expected to create a subdirectory at the input path inside dir. All files (artifacts and metadata documents) required to represent x on disk should be created inside path. The expected contents of path are:

Exactly one JSON metadata document with a name ending in .json. This contains the metadata describing x, according to the schema defined in the $schema property. If a data file exists (see next), the path property should contain the relative path to that file from dir; otherwise it should contain the relative path of the metadata document itself.

(Optional) A file containing the data inside x. This file should have the same name as the metadata file after stripping the .json extension. Methods are free to choose any format and name within path except for the .json file extension, which is reserved for metadata documents created with .writeMetadata.

(Optional) Further subdirectories containing child objects of x. Each child object should be saved in its own subdirectory, and should be referenced via a resource object within the metadata for x. When saving children, methods should call .stageObject rather than stageObject (note the period at the start of the former). This ensures that the staging method will respect customizations from alabaster applications that define their own generic in .altStageObject.

Methods can create both a data file and multiple subdirectories. In this manner, we can decompose complex x into their components for easier handling.

Keep in mind that the returned path will differ from the input path; the latter refers to the directory while the former refers to a file inside the directory. Methods should use the former to reference x from the metadata of a parent object.

The stageObject generic will check if the path already exists before dispatching to the methods. If so, it will throw an error to ensure that downstream name clashes do not occur. The exception is if path = ".", in which case no check is performed; this is useful for eliminating subdirectories in situations where the project contains only one object.

Any attempt to use the stageObject generic to save another non-child object into path or its subdirectories will cause an error. This ensures that path contains all and only the contents of x.

See also

Author

Aaron Lun

Examples

tmp <- tempfile()
dir.create(tmp)

library(S4Vectors)
X <- DataFrame(X=LETTERS, Y=sample(3, 26, replace=TRUE))
stageObject(X, tmp, path="test1")
#> $`$schema`
#> [1] "csv_data_frame/v1.json"
#> 
#> $path
#> [1] "test1/simple.csv.gz"
#> 
#> $is_child
#> [1] FALSE
#> 
#> $data_frame
#> $data_frame$columns
#> $data_frame$columns[[1]]
#> $data_frame$columns[[1]]$name
#> [1] "X"
#> 
#> $data_frame$columns[[1]]$type
#> [1] "string"
#> 
#> 
#> $data_frame$columns[[2]]
#> $data_frame$columns[[2]]$name
#> [1] "Y"
#> 
#> $data_frame$columns[[2]]$type
#> [1] "integer"
#> 
#> 
#> 
#> $data_frame$row_names
#> [1] FALSE
#> 
#> $data_frame$column_data
#> NULL
#> 
#> $data_frame$other_data
#> NULL
#> 
#> $data_frame$dimensions
#> [1] 26  2
#> 
#> 
#> $csv_data_frame
#> $csv_data_frame$compression
#> [1] "gzip"
#> 
#> 
list.files(file.path(tmp, "test1"))
#> [1] "simple.csv.gz"